hacker_news_reskin/requirements.txt
openai
python-fasthtml
aiohttp
beautifulsoup4